Add stop(), Add trackedPublish(), move globals
Add a stop() function which is called to gracefully disconnect without raising an error using a new global shouldBeConnected.

Add a trackedPublish() wrapper to client.publish() which adds the message id to a global messagesInTransit so that the messsage can be handled in a new call back handle_publish()

Add MQTTDisconnectError to be raised if unintentional disconnect occurs.