Change test for last will to socket
Remove need for human Ctrl-C to test that last will works. Instead call the shutdown() method of the MQTT client's socket to simulate an ungraceful disconnect.

Remove the error raised by the dummy Smart Agent class in test_provider when there is an unexpected shutdown. Also remove the variable tracking this.