Add TOPIC_REQUEST
Add a topic for broker-services to post on when it starts up. When SmartAgents recieve a message on this topic, they must repost their connected status so that broker-services registers them.

Removed TODO message relating to how to deal with retained discovery from previous broker-services connection. This is not a problem with the new topic and SmartAgents refreshing their status.

Change setting global 'connected' to after checking if there is an error.

Add test for sending this request message on a broker-services startup. This includes adding a list of messages recieved by the dummy SmartAgent class.